[#2] Fix MMC1 mapper (#23)
* [#2] Fixed some logs

* [#2] Switched mapped to return usize and fix mapper 1 RAM write

* [#2] Small fixes

+++ server/nestadia-core/src/cartridge/mapper_002.rs
@@ -19,8 +19,8 @@ impl Mapper002 {
 impl Mapper for Mapper002 {
     fn cpu_map_read(&self, addr: u16) -> CartridgeReadTarget {
         match addr {
-            0x8000..=0xBFFF => CartridgeReadTarget::PrgRom((self.prg_bank_selector as u16) * 0x4000 + (addr & 0x3FFF)),
-            _ => CartridgeReadTarget::PrgRom((self.prg_banks as u16 - 1) * 0x4000 + (addr & 0x3FFF)),
+            0x8000..=0xBFFF => CartridgeReadTarget::PrgRom((self.prg_bank_selector as usize) * 0x4000 + (addr & 0x3FFF) as usize),
+            _ => CartridgeReadTarget::PrgRom((self.prg_banks as usize - 1) * 0x4000 + (addr & 0x3FFF) as usize),
         }
     }
 
@@ -28,11 +28,11 @@ impl Mapper for Mapper002 {
         self.prg_bank_selector = data;
     }
 
-    fn ppu_map_read(&self, addr: u16) -> u16 {
-        addr
+    fn ppu_map_read(&self, addr: u16) -> usize {
+        addr as usize
     }
 
-    fn ppu_map_write(&self, _addr: u16) -> Option<u16> {
+    fn ppu_map_write(&self, _addr: u16) -> Option<usize> {
         None
     }
